seperate HTML extractors based on the HTML structure
