+++ src/main.rb
@@ -1,7 +1,8 @@
 require_relative "GoogleAppBarExtractor"
 
-html = File.read(__dir__ + "/../files/mu-fc-players.html")
+html = File.read(__dir__ + "/tests/mu-fc-players.html")
 
 extractor = GoogleAppBarExtractor.new
 
-File.write(__dir__ + '/res.json' , extractor.extract_cards(html).to_json)
+extractor.extract_cards(html)
+# File.write(__dir__ + '/res.json' , extractor.extract_cards(html).to_json)
